move gcov.h to sys/gnu/gcov
